moved helper method extract_paths_from_metalink to tests/common.py
